requirements.txt
appdirs==1.4.0
CouchDB==1.1
Django==1.10.5
django-filter==1.0.1
django-nose==1.4.4
django-redis==4.7.0
djangorestframework==3.5.3
fakeredis==0.8.2
future==0.16.0
gunicorn==19.6.0
httplib2==0.10.3
magicmock==0.3
Markdown==2.6.8
mock==2.0.0
nose==1.3.7
packaging==16.8
pbr==1.10.0
pyparsing==2.1.10
python-memcached==1.58
redis==2.10.5
requests==2.13.0
six==1.10.0
